+++ server/src/GemLevelProtScraper/Program.cs
@@ -103,9 +103,9 @@ app
         [FromServices] PoeRepository poeRepository,
         [FromQuery(Name = "league")] string league = "",
         [FromQuery(Name = "gem_name")] string gemNameWindcard = "*",
-        [FromQuery(Name = "min_sell_price_chaos")] decimal? minSellPriceChaos = null,
-        [FromQuery(Name = "max_buy_price_chaos")] decimal? maxBuyPriceChaos = null,
-        [FromQuery(Name = "min_experience_delta")] decimal? minExperienceDelta = null,
+        [FromQuery(Name = "min_sell_price_chaos")] double? minSellPriceChaos = null,
+        [FromQuery(Name = "max_buy_price_chaos")] double? maxBuyPriceChaos = null,
+        [FromQuery(Name = "min_experience_delta")] double? minExperienceDelta = null,
         [FromQuery(Name = "min_listing_count")] int minListingCount = 4,
         [FromQuery(Name = "items_count")] int itemsCount = 10,
         CancellationToken cancellationToken = default
